requirements.txt
biopython==1.79
boto3==1.22.7
botocore==1.25.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
docker-pycreds==0.4.0
fair-esm==0.4.2
filelock==3.6.0
gitdb==4.0.9
GitPython==3.1.27
huggingface-hub==0.5.1
idna==3.3
jmespath==1.0.0
joblib==1.1.0
lmdb==1.3.0
mkl-fft==1.3.1
mkl-service==2.4.0
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pathtools==0.1.2
promise==2.3
protobuf==3.20.1
psutil==5.9.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
s3transfer==0.5.2
sacremoses==0.0.53
scikit-learn==1.0.2
scipy==1.8.0
sentry-sdk==1.5.12
seqeval==1.2.2
setproctitle==1.2.3
shortuuid==1.0.9
smmap==5.0.0
tape-proteins==0.5
tensorboardX==2.5
threadpoolctl==3.1.0
tokenizers==0.10.3
torch==1.11.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
wandb==0.12.16
